+++ app/src/main/java/ru/art241111/gt_kawasaki/repository/robotAPI/RobotEntity.kt
@@ -18,5 +18,5 @@ class RobotEntity {
     val errors: Queue<String> = LinkedList<String>()
     var state: State = State.WAITING_COMMAND
 
-    var position: MutableList<Float> = mutableListOf()
+    var position: MutableList<Double> = mutableListOf()
 }
